LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/reload-intercept-reject_no-currententrychange-expected.txt


FAIL event and promise ordering for navigation.reload() intercepted by intercept() assert_array_equals: expected property 2 to be "AbortSignal abort" but got "committed fulfilled" (expected array ["navigate", "handler run", "AbortSignal abort", "navigateerror", "committed fulfilled", "promise microtask", "finished rejected", "transition.finished rejected"] got ["navigate", "handler run", "committed fulfilled", "promise microtask", "AbortSignal abort", "navigateerror", "finished rejected", "transition.finished rejected"])
